add oauth2 authentication for routes

+++ project/app/routers/reviews.py
@@ -1,8 +1,10 @@
-from fastapi import HTTPException, APIRouter
+from fastapi import HTTPException, APIRouter, Depends
 
 from ..database import UserReview, User, Movie
 
 from ..schemas import ReviewRequestModel, ReviewResponseModel, ReviewRequestPutModel
+
+from ..common import get_current_user
 
 from typing import List
 
@@ -10,16 +12,13 @@ router = APIRouter(prefix='/reviews')
 
 
 @router.post('', response_model=ReviewResponseModel)
-async def create_reviews(user_review: ReviewRequestModel):
-
-    if User.select().where(User.id == user_review.user_id).first() is None:
-        raise HTTPException(status_code=404, detail='Usuario no encontrado.')
+async def create_reviews(user_review: ReviewRequestModel, user: User = Depends(get_current_user)):
 
     if Movie.select().where(Movie.id == user_review.movie_id).first() is None:
         raise HTTPException(status_code=404, detail='Pelicula no encontrada.')
 
     user_review = UserReview.create(
-        user=user_review.user_id,
+        user=user.id,
         movie=user_review.movie_id,
         review=user_review.review,
         score=user_review.score
@@ -45,12 +44,15 @@ async def get_review(review_id: int):
 
 
 @router.put('/{review_id}', response_model=ReviewResponseModel)
-async def update_review(review_id: int, review_request: ReviewRequestPutModel):
+async def update_review(review_id: int, review_request: ReviewRequestPutModel, user: User = Depends(get_current_user)):
 
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(status_code=404, detail='Review no encontrada.')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(status_code=401, detail='Esta no es tu reseña.')
 
     user_review.review = review_request.review
     user_review.score = review_request.score
@@ -60,11 +62,15 @@ async def update_review(review_id: int, review_request: ReviewRequestPutModel):
 
 
 @router.delete('/{review_id}', response_model=ReviewResponseModel)
-async def delete_review(review_id: int):
+async def delete_review(review_id: int, user: User = Depends(get_current_user)):
+    
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(status_code=404, detail='Review no encontrada.')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(status_code=401, detail='Esta no es tu reseña.')
 
     user_review.delete_instance()
